requirements.txt
pytest-django
pytest-mock
factory-boy
sqlparse
